refactor: mirror test layout with source structure
Organize tests under paths matching their runtime modules so related source and coverage are easier to locate.
