+++ docs/adr/0023-url-filters-accept-a-subset-of-chrome-match-patterns.md
@@ -28,4 +28,4 @@ Alternatives rejected:
 - The query part is matched on its own, so `https://www.google.com/search?*` also matches the bare `/search`. Chrome would not. Lenient only.
 - Every entry that worked in 9.1.1 keeps working. The rejected shapes broke Chrome's registration or failed the old validator, so no synced entry of those shapes ever filtered a page.
 - The popup status row and the icon stop mislabeling `*.` entries, since they share the function.
-- `tests/browser/urlpattern.playwright.ts` keeps testing raw `URLPattern`, including the apex mismatch. It records why the translation exists.
+- `tests/extension/settings/urlpattern.playwright.ts` keeps testing raw `URLPattern`, including the apex mismatch. It records why the translation exists.
