+++ playwright.config.ts
@@ -4,7 +4,7 @@ import { defineConfig, devices } from '@playwright/test';
  * See https://playwright.dev/docs/test-configuration.
  */
 export default defineConfig({
-  testDir: './tests/browser',
+  testDir: './tests',
   testMatch: '**/*.playwright.ts',
   /* Run tests in files in parallel */
   fullyParallel: true,
