+++ tests/browser/dom/visibility-detector.playwright.ts
