+++ tests/extension/ai-spacing/ai-spacing.test.ts
@@ -1,7 +1,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
-import type { ClassifyCandidatesMessage } from '../../browser-extensions/chrome/src/ai-spacing/messages';
-import type { LateFix } from '../../src/browser/pangu';
-import { pangu as corePangu } from '../../src/shared/index';
+import type { ClassifyCandidatesMessage } from '../../../browser-extensions/chrome/src/ai-spacing/messages';
+import type { LateFix } from '../../../src/browser/pangu';
+import { pangu as corePangu } from '../../../src/shared/index';
 
 async function loadAiSpacing() {
   vi.resetModules();
@@ -13,11 +13,11 @@ async function loadAiSpacing() {
       }
     }),
   };
-  vi.doMock('../../src/browser/pangu', () => ({ default: pangu }));
-  const { handleClassification } = await import('../../browser-extensions/chrome/src/ai-spacing/service-worker');
+  vi.doMock('../../../src/browser/pangu', () => ({ default: pangu }));
+  const { handleClassification } = await import('../../../browser-extensions/chrome/src/ai-spacing/service-worker');
   const sendMessage = vi.fn(({ kind, candidates }: ClassifyCandidatesMessage) => handleClassification(kind, candidates));
   vi.stubGlobal('chrome', { runtime: { sendMessage } });
-  const { applyAiSpacing, warmUpAiSpacing } = await import('../../browser-extensions/chrome/src/ai-spacing/content-script');
+  const { applyAiSpacing, warmUpAiSpacing } = await import('../../../browser-extensions/chrome/src/ai-spacing/content-script');
   async function spaceTextWithAi(unspaced: string) {
     const settled = corePangu.spaceText(unspaced);
     const node = { data: settled } as Text;
